16 -> 32 bit network packet lengths... NO BACKWARD COMPATABILITY

+++ include/liblattice/lattice_packet.h
@@ -31,40 +31,40 @@
 
   // --------------------
 
-  extern int get_int8(void **src, int8_t *dst, uint16_t *len, uint16_t *argc);
-  extern int get_int16(void **src, int16_t *dst, uint16_t *len, uint16_t *argc);
-  extern int get_int32(void **src, int32_t *dst, uint16_t *len, uint16_t *argc);
-  extern int get_int64(void **src, int64_t *dst, uint16_t *len, uint16_t *argc);
+  extern int get_int8(void **src, int8_t *dst, uint32_t *len, uint32_t *argc);
+  extern int get_int16(void **src, int16_t *dst, uint32_t *len, uint32_t *argc);
+  extern int get_int32(void **src, int32_t *dst, uint32_t *len, uint32_t *argc);
+  extern int get_int64(void **src, int64_t *dst, uint32_t *len, uint32_t *argc);
 
-  extern int get_uint8(void **src, uint8_t *dst, uint16_t *len, uint16_t *argc);
-  extern int get_uint16(void **src, uint16_t *dst, uint16_t *len, uint16_t *argc);
-  extern int get_uint32(void **src, uint32_t *dst, uint16_t *len, uint16_t *argc);
-  extern int get_uint64(void **src, uint64_t *dst, uint16_t *len, uint16_t *argc);
+  extern int get_uint8(void **src, uint8_t *dst, uint32_t *len, uint32_t *argc);
+  extern int get_uint16(void **src, uint16_t *dst, uint32_t *len, uint32_t *argc);
+  extern int get_uint32(void **src, uint32_t *dst, uint32_t *len, uint32_t *argc);
+  extern int get_uint64(void **src, uint64_t *dst, uint32_t *len, uint32_t *argc);
 
-  extern int get_float(void **src, float *dst, uint16_t *len, uint16_t *argc);
-  extern int get_double(void **src, double *dst, uint16_t *len, uint16_t *argc);
-  extern int get_string(void **src, char **dst, uint16_t *len, uint16_t *argc);
+  extern int get_float(void **src, float *dst, uint32_t *len, uint32_t *argc);
+  extern int get_double(void **src, double *dst, uint32_t *len, uint32_t *argc);
+  extern int get_string(void **src, char **dst, uint32_t *len, uint32_t *argc);
 
-  extern int get_sector(void **src, block_t *dst, uint16_t *len, uint16_t *argc);
-  extern int get_flatsector(void **src, flatdata_t *dst, uint16_t *len, uint16_t *argc);
+  extern int get_sector(void **src, block_t *dst, uint32_t *len, uint32_t *argc);
+  extern int get_flatsector(void **src, flatdata_t *dst, uint32_t *len, uint32_t *argc);
 
   // --------------------
 
-  extern int put_int8(void **dst, int8_t src, uint16_t *len, uint16_t *argc);
-  extern int put_int16(void **dst, int16_t src, uint16_t *len, uint16_t *argc);
-  extern int put_int32(void **dst, int32_t src, uint16_t *len, uint16_t *argc);
-  extern int put_int64(void **dst, int64_t src, uint16_t *len, uint16_t *argc);
+  extern int put_int8(void **dst, int8_t src, uint32_t *len, uint32_t *argc);
+  extern int put_int16(void **dst, int16_t src, uint32_t *len, uint32_t *argc);
+  extern int put_int32(void **dst, int32_t src, uint32_t *len, uint32_t *argc);
+  extern int put_int64(void **dst, int64_t src, uint32_t *len, uint32_t *argc);
 
-  extern int put_uint8(void **dst, uint8_t src, uint16_t *len, uint16_t *argc);
-  extern int put_uint16(void **dst, uint16_t src, uint16_t *len, uint16_t *argc);
-  extern int put_uint32(void **dst, uint32_t src, uint16_t *len, uint16_t *argc);
-  extern int put_uint64(void **dst, uint64_t src, uint16_t *len, uint16_t *argc);
+  extern int put_uint8(void **dst, uint8_t src, uint32_t *len, uint32_t *argc);
+  extern int put_uint16(void **dst, uint16_t src, uint32_t *len, uint32_t *argc);
+  extern int put_uint32(void **dst, uint32_t src, uint32_t *len, uint32_t *argc);
+  extern int put_uint64(void **dst, uint64_t src, uint32_t *len, uint32_t *argc);
 
-  extern int put_float(void **dst, float src, uint16_t *len, uint16_t *argc);
-  extern int put_double(void **dst, double src, uint16_t *len, uint16_t *argc);
-  extern int put_string(void **dst, const char *src, uint16_t *len, uint16_t *argc);
+  extern int put_float(void **dst, float src, uint32_t *len, uint32_t *argc);
+  extern int put_double(void **dst, double src, uint32_t *len, uint32_t *argc);
+  extern int put_string(void **dst, const char *src, uint32_t *len, uint32_t *argc);
 
-  extern int put_sector(void **dst, block_t *src, uint16_t *len, uint16_t *argc);
-  extern int put_flatsector(void **dst, flatdata_t *src, uint16_t *len, uint16_t *argc);
+  extern int put_sector(void **dst, block_t *src, uint32_t *len, uint32_t *argc);
+  extern int put_flatsector(void **dst, flatdata_t *src, uint32_t *len, uint32_t *argc);
 
 #endif
